fixed test util
